requirements.txt
pandas
bertopic
nltk
sentence-transformers
polars
python-dotenv